DOC: fix the copybutton on the code blocks (#7634)
The copybutton is the ">>>" on the side of code blocks that enables to
remove the ">>>" at the begining of lines as well as the outputs.

It was broken in the docs of 0.18 due to changes in sphinx. I am
backporting the js used by the Python core docs.

In addition, I am backporting CSS to highlight this button
on hover.